refactor: restructure almost all files in src/ui
