+++ src/pages/Panel/index.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import Panel from '../../ui/index';
+import Devtools from '../../ui/index';
 
-render(<Panel />, window.document.querySelector('#panel'));
+render(<Devtools />, window.document.querySelector('#panel'));
 
 if (module.hot) module.hot.accept();
