+++ src/ui/NetSpeed/NetSpeed.scss
@@ -14,14 +14,14 @@
 }
 
 .icon-upload {
-  background: url('./buttonIcons/upload.svg');
+  background: url('../../assets/icons/upload.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-download {
-  background: url('./buttonIcons/download.svg');
+  background: url('../../assets/icons/download.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
