+++ src/ui/NoMessageSelected/NoMessageSelected.scss
