+++ src/ui/ResizableTable/ResizableTable.scss
