+++ src/ui/Rpc/Rpc.scss
@@ -75,56 +75,56 @@
 }
 
 .icon-record {
-  background: url('./buttonIcons/record.svg');
+  background: url('../../assets/icons/record.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-clear {
-  background: url('./buttonIcons/clear.svg');
+  background: url('../../assets/icons/clear.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-bottom {
-  background: url('./buttonIcons/bottom.svg');
+  background: url('../../assets/icons/bottom.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-filter {
-  background: url('./buttonIcons/filter.svg');
+  background: url('../../assets/icons/filter.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-reset {
-  background: url('./buttonIcons/reset.svg');
+  background: url('../../assets/icons/reset.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-expand-all {
-  background: url('./buttonIcons/expand_all.svg');
+  background: url('../../assets/icons/expand_all.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-collapse-all {
-  background: url('./buttonIcons/collapse_all.svg');
+  background: url('../../assets/icons/collapse_all.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
 }
 
 .icon-braces {
-  background: url('./buttonIcons/braces.svg');
+  background: url('../../assets/icons/braces.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center center;
