+++ webpack.config.js
@@ -135,7 +135,7 @@ let options = {
     new CopyWebpackPlugin({
       patterns: [
         {
-          from: 'src/assets/img/logo.png',
+          from: 'src/assets/logo.png',
           to: path.join(__dirname, 'build'),
           force: true,
         },
